Remove broken OpenAPI validation
This workflow attempts to validate the openAPI.yaml file,
but it never passed, not even at the time it was first added.

The configuration it uses for the linter is now outdated - the linter no longer uses .validaterc.

Modifying the spec to adhere to what the linter expects would,
at this point, involve modifying the interface exposed by other projects,
which would be a significant undertaking.

Additonally, the API spec hasn't been modified in years, so linting it
regularly doesn't seem particularly useful.